Problem: understanding programs is difficult
Literate programming helps to a degree by establishing
a train of thought, together with thorough explanation,
of the developers behind the program.

However, it is very text chunk oriented, meaning there's
very little to derive from it besides the explanation
and the flow of thought. Not to undervalue this benefit,
it still forces us to keep declare all the minutiae of
writing a program, such as keeping an appropriate file
structure which forces us to split out the chunks in
way most appropriate for such structuring.

Besides, it doesn't allow us to do any conceptual
introspection of a program.

Solution: start an experimental "construct macro"
expansion technique.

The idea behind it is to define the entire program
*and* the associated literature as a body of knowledge
facts ("constructs") and use a rule production system
to infer supplemental knowledge and produce verbatim files
("macro expansion").

This technique also allows to browse constructs defined
in the document without any expansion, and define custom
rules and construct templates that would allow us to
express a nearly unlimited variety of of concepts.

This inital code import contains a minimalistic self-hosted
implementation of a tiny core for construct macro expansion.

It is deliberately limited as the goal was to bootstrap it
through itself. It does not currenly provide any constructs
for literature, for example. The design decisions are temporary
at best as the plan is to iterate from here and find what works
best.

To expand `conmacro.clp` into `conmacro_boot.clp`, run:

```
clips -f2 conmacro_boot.clp -f2 conmacro.clp
```

Let's see how far can we take this.