Problem: Projects without README.md are less welcoming
Solution: produce one

As an exercise, instead of writing a static one, this
README.md was defined through conmacro.clp. Not particularly
useful at this moment, but it allowed to slightly evolve
the file production system by allowing to register
mappings between file extensions and types (`file-extension`
construct) and, more importantly, it forced us to
introduce `clips:assertion` construct that allows to produce
an assertion.