Problem: CLIPS syntax is overwhelmingly verbose
I am drowning in the sea of assertions and parens
and it is hard to make out what exactly I am writing.

This slows development down a lot.

Solution: design a preliminary simplified syntax
called Unobtrusive Construct Syntax (file name
extenion .ucon). It is not the best name, but
I was sleep-deprived enough for this. Whenever
a better name will come, we can rename it.

The idea behind the language is to make it easy to
read and support some extra features CLIPS doesn't
(unindented multiline strings, for example or special
shortcuts for common attributes)

For the time being I dropped the markdown generation
part just to keep the core smaller.

And yes, uconsyn tool hasn't been built using conmacro
just yet. This is because I didn't want to drown in
the original syntax too early.